backend/requirements.txt
alembic==1.16.4
apispec==6.8.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==25.3.0
bidict==0.23.1
black==25.1.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
dnspython==2.8.0
eventlet==0.35.2
flasgger==0.9.7.1
Flask==3.1.1
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-Session==0.8.0
flask-smorest==0.46.1
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
minio==7.2.16
mistune==3.1.3
msgspec==0.19.0
mypy_extensions==1.1.0
netifaces==0.10.6
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.23.0
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
ruff==0.12.3
setuptools==3.3
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.1
urllib3==2.5.0
webargs==8.7.0
Werkzeug==3.1.3
wsproto==1.2.0
PyJWT==2.8.0
